+++ graphAnalyser/metrics/metric.h
@@ -8,7 +8,7 @@ class Metric {
     public:
         Metric(std::string name) : name(name) {
         }
-        virtual int calculate(const Graph& graph) const = 0;
+        virtual double calculate(const Graph& graph) const = 0;
         const std::string name;
 };
 #endif
